requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
bidict==0.23.1
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
Faker==37.1.0
Flask==3.1.0
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limiter==0.5.0
limits==4.7.2
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.1
pyotp==2.9.0
python-dotenv==1.1.0
python-engineio==4.11.2
python-socketio==5.12.1
qrcode==8.2
redis==6.0.0
rich==13.9.4
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
StrEnum==0.4.15
token-bucket==0.3.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
tzlocal==5.3.1
uvicorn==0.34.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
